moved files into beamVisualizer deleted extras
