--- README.md
@@ -1,1 +1,0 @@
-beam-project-readme
