+++ beamVisualizer/client/styles/globals.css
