+++ beamVisualizer/client/trashComponents/ChoiceBox.jsx
