+++ beamVisualizer/client/trashComponents/Graph.jsx
