+++ beamVisualizer/client/trashComponents/Navigation.jsx
