+++ beamVisualizer/client/trashComponents/PositionsTable.jsx
