+++ beamVisualizer/server/beamServer/controllers/gqlController.ts
